+++ CHANGES.rst
@@ -5,6 +5,7 @@ Changes in version 1.1 (in development)
 ---------------------------------------
 
 * Fix checking ``unique_together`` on the translated model.
+* Added ``SortedRelatedFieldListFilter`` for displaying translated models in the ``list_filter``.
 
 
 Changes in version 1.0 (2014-07-07)